MDL-22787 MDL-21993 New mnetservice plugin type
This commit introduces new plugin type called mnetservice. Once the MNet
framework is redesigned, this plugin will define the service itself, the
strings for it and the the client tools for administrator. The server
side is implemented mainly in enrol_mnet plugin which publishes xml-rpc
methods.

AMOS BEGIN
 MOV [remoteenrolhosts_desc,core_mnet],[clientname_help,mnetservice_enrol]
AMOS END